requirement.txt
transformers-4.34.0
bitsandbytes-0.41.1
peft-0.5.0
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
datasets-2.14.5
pip install fschat==0.1.2
scipy-1.11.3
accelerate-0.23.0
coloredlogs-15.0.1 humanfriendly-10.0 optimum-1.13.2